Coach Animation for Bump
Adjustments to current animationTime to specify name for jump animation. Added bump keyboard control.
Not really thrilled with having to navigate the hierarchy to get the end of the left hand to have a ball location. At least it's cached. But also was not thrilled about navigating the current structure with the name "Volleyball-Male." Change the name to generic "Character." This works out since the graphics are below that and enabled separately.