Improved Spike process and parabolic curve formula
Thanks to Claude, was able to change the formula for the curve to exactly what I wanted!
Thanks to discussion with Robert, will look to see if there's a direct line to the target - if there is then spike direct, otherwise use a pass with a max height of the starting height to make it appear more direct but still get over the net.