More on Match Administration Setup - debug
Tweaked the MatchInfoSO to display teams in the inspector. TeamSO adjusted to be read only with a function to add (and clear in Init). Created debug scripts to add an actor already in the scene to a TeamSO and another to clear the TeamSO (to initialize).

Looks like this is working well and I might finally have it the way I want/need!