Tested instantiation of AI Prefabs
Had to do some custom debug stuff to test it out, but appears to be working right. Having the loading of the teams on a prior stage makes for testing the MatchManager in the same scene difficult. Hopefully this works well in an actual game situation (configure match scene, then game scene), so will really need to work on completing that next, which will be a fully working game! 😬
Also set some default eyes/hair/skin for prefabs and a note in the future updates to add this customization.