Refactored Outfit Menu UI selection to use the Enum
Since this is the only "manual" one, this has been changed to use the existing Enum. This means it's not modular (did look into using an Enum selector), but I think it's fine for this use case. Also made a tweak to resetting the index when updating from a parent change.