Began work on Adjusting AI Serve for Net Collisions (tFLJxJLW)
Lots of tests for how to try to predict if the serve will be over the net or not.  Finally came up with a good solution and have created a checklist of approach for refactoring the code to apply the new logic.  This commit begins some of that with the prediction code as well as moving the ball flight position to the ball SO for more broad access.