Spaced out AI after first hit for better Spike position (tsCDrWXV)
Change OffensePos property to instead be in the OffenseState, set once, with randomness and a buffer.  This is the only place it was used, so this makes sense to remove the property.
Also some residual stuff that should have been in the last commit like an Excel calculating Dot products and Angles and commenting some debug logs