Adjustment to CoachAction to ensure Coach has the Ball
See the commentary on the BumpCoach class.  It's using the same code in its action as the AI DigReadyState.  Should be DRY.  But the only other place it's used is that class (by the AI).  Therefore, it could be moved up a level to the AI but then the Coach would need to inherit from AI and not from Athlete.  This would make sense (if it needs to inherit from a derived class of Athlete, it's much closer to AI than Player, of course), but is there any AI functionality that is inherent that is not desired for Coach?

Also, ensuring the Coach has the Ball added a boolean to check for it but most notably, a revision to the function that the inherited classes need to override.