Getting profile to load from separate setup script (hjHlRz1Y)
Side note:  this ProjectSettings change was new - not sure why that occurred.  Also, there was a bin and obj folder and many subfiles added to the Scripts\Actors folder that resulted in compilation errors:  Global statements needed .NET 10 and it was in .NET 9.  Not sure why this downgrade occurred since I just booted up for the day and it was working fine last night when I loaded.